+++ apps/astarte_realm_management_api/test/support/astarte_realm_management_mock.ex
@@ -16,10 +16,17 @@ defmodule Astarte.RealmManagement.Mock do
     InstallInterface,
     Reply,
     UpdateInterface,
-    UpdateJWTPublicKeyPEM
+    UpdateJWTPublicKeyPEM,
+    InstallTriggerPolicy,
+    GetTriggerPoliciesList,
+    GetTriggerPoliciesListReply,
+    DeleteTriggerPolicy,
+    GetTriggerPolicySource,
+    GetTriggerPolicySourceReply
   }
 
   alias Astarte.Core.Interface
+  alias Astarte.Core.Triggers.Policy
   alias Astarte.RealmManagement.Mock.DB
 
   def rpc_call(payload, _destination) do
@@ -153,6 +160,72 @@ defmodule Astarte.RealmManagement.Mock do
 
     generic_ok()
     |> ok_wrap
+  end
+
+  defp execute_rpc(
+         {:install_trigger_policy,
+          %InstallTriggerPolicy{realm_name: realm_name, trigger_policy_json: trigger_policy_json}}
+       ) do
+    {:ok, params} = Jason.decode(trigger_policy_json)
+
+    {:ok, policy} = Policy.changeset(%Policy{}, params) |> Ecto.Changeset.apply_action(:insert)
+
+    with :ok <- DB.install_trigger_policy(realm_name, policy) do
+      generic_ok(true)
+      |> ok_wrap
+    else
+      {:error, reason} ->
+        generic_error(reason)
+        |> ok_wrap
+    end
+  end
+
+  defp execute_rpc(
+         {:get_trigger_policies_list,
+          %GetTriggerPoliciesList{
+            realm_name: realm_name
+          }}
+       ) do
+    list = DB.get_trigger_policies_list(realm_name)
+
+    %GetTriggerPoliciesListReply{trigger_policies_names: list}
+    |> encode_reply(:get_trigger_policies_list_reply)
+    |> ok_wrap
+  end
+
+  defp execute_rpc(
+         {:delete_trigger_policy,
+          %DeleteTriggerPolicy{
+            realm_name: realm_name,
+            trigger_policy_name: name
+          }}
+       ) do
+    case DB.delete_trigger_policy(realm_name, name) do
+      :ok ->
+        generic_ok()
+        |> ok_wrap()
+
+      {:error, reason} ->
+        generic_error(reason)
+        |> ok_wrap()
+    end
+  end
+
+  defp execute_rpc(
+         {:get_trigger_policy_source,
+          %GetTriggerPolicySource{
+            realm_name: realm_name,
+            trigger_policy_name: name
+          }}
+       ) do
+    if source = DB.get_trigger_policy_source(realm_name, name) do
+      %GetTriggerPolicySourceReply{source: source}
+      |> encode_reply(:get_trigger_policy_source_reply)
+      |> ok_wrap
+    else
+      generic_error(:trigger_policy_not_found)
+      |> ok_wrap
+    end
   end
 
   defp generic_ok(async_operation \\ false) do